wip: Initial slick carousel implementation
Ultimately, the actual styling required for a Carousel is minimal — mostly just the navigation dots and/or pager arrows. Slick supports custom components for those elements. The bigger issue with Carousels is the range of options required for structural behaviour of the Carousel — swiping, navigation, slides-per-page, responsive behaviour, etc. And on that front, Slick seems to be as good as any.