feat: Revert to use styled-components (instead of Emotion styled)
Dunno why, but the Styled-Components version of styled function seems to work correctly, when extending components. Emotion doesn’t work properly (in most cases).
Need to keep an eye on this, as per Rebass doco (somewhere) there are inconsistencies in how Emotion, Styled-Components and themes work together.
However … so far, so good.